chess-ai-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-chess==1.2.0
torch==2.1.0
numpy==1.24.3
pydantic==2.4.2
python-multipart==0.0.6